+++ source/common/stats/histogram_impl.h
@@ -36,13 +36,13 @@ public:
   const std::vector<double>& supportedQuantiles() const override;
   const std::vector<double>& computedQuantiles() const override { return computed_quantiles_; }
   const std::vector<double>& supportedBuckets() const override;
-  const std::vector<double>& computedBuckets() const override { return computed_buckets_; }
+  const std::vector<uint64_t>& computedBuckets() const override { return computed_buckets_; }
   double sampleCount() const override { return sample_count_; }
   double sampleSum() const override { return sample_sum_; }
 
 private:
   std::vector<double> computed_quantiles_;
-  std::vector<double> computed_buckets_;
+  std::vector<uint64_t> computed_buckets_;
   double sample_count_;
   double sample_sum_;
 };
